refactor specification zome files into new module structure
no code changes, except for struct field visibility
