+++ happs/specification/zomes/action/code/Cargo.toml
@@ -6,12 +6,8 @@ edition = "2018"
 
 [dependencies]
 serde = "=1.0.89"
-serde_json = { version = "=1.0.39", features = ["preserve_order"] }
-serde_derive = "=1.0.89"
 hdk = "=0.0.40-alpha1"
 hdk_proc_macros = "=0.0.40-alpha1"
-holochain_wasm_utils = "=0.0.40-alpha1"
-holochain_json_derive = "=0.0.17"
 
 vf_knowledge = { path = "../../../../../lib/vf_knowledge" }
 vf_core = { path = "../../../../../lib/vf_core" }
