+++ happs/specification/zomes/process_specification/code/Cargo.toml
@@ -15,7 +15,6 @@ hdk_proc_macros = "0.0.40-alpha1"
 vf_core = { path = "../../../../../lib/vf_core" }
 hc_zome_rea_process_specification_defs = { path = "../../../../../lib/rea_process_specification/defs" }
 hc_zome_rea_process_specification_storage_consts = { path = "../../../../../lib/rea_process_specification/storage_consts" }
-hc_zome_rea_process_specification_storage = { path = "../../../../../lib/rea_process_specification/storage" }
 hc_zome_rea_process_specification_rpc = { path = "../../../../../lib/rea_process_specification/rpc" }
 hc_zome_rea_process_specification_lib = { path = "../../../../../lib/rea_process_specification/lib" }
