+++ happs/specification/zomes/unit/code/Cargo.toml
@@ -15,7 +15,6 @@ hdk_proc_macros = "0.0.40-alpha1"
 vf_core = { path = "../../../../../lib/vf_core" }
 hc_zome_rea_unit_defs = { path = "../../../../../lib/rea_unit/defs" }
 hc_zome_rea_unit_storage_consts = { path = "../../../../../lib/rea_unit/storage_consts" }
-hc_zome_rea_unit_storage = { path = "../../../../../lib/rea_unit/storage" }
 hc_zome_rea_unit_rpc = { path = "../../../../../lib/rea_unit/rpc" }
 hc_zome_rea_unit_lib = { path = "../../../../../lib/rea_unit/lib" }
