--- happs/specification_DEPRECATED/.hcignore
@@ -1,3 +1,0 @@
-dist
-test
-README.md
