--- happs/specification_DEPRECATED/app.json
@@ -1,19 +1,0 @@
-{
-  "name": "HoloREA: specification",
-  "description": "Specification DNA for the REA economic network",
-  "authors": [
-    {
-      "identifier": "JhonatanHern <jhonatanhernandez998@gmail.com>",
-      "public_key_source": "",
-      "signature": ""
-    },
-    {
-      "identifier": "pospi <pospi@spadgos.com>",
-      "public_key_source": "",
-      "signature": ""
-    }
-  ],
-  "version": "0.1.0",
-  "dht": {},
-  "properties": null
-}
