--- happs/specification_DEPRECATED/zomes/resource_specification/code/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "hc_zome_rea_specification_resource_specification"
-version = "0.1.0"
-authors = ["JhonatanHern"]
-edition = "2018"
-
-[dependencies]
-serde = "=1.0.89"
-serde_json = { version = "=1.0.39", features = ["preserve_order"] }
-serde_derive = "=1.0.89"
-hdk = "=0.0.40-alpha1"
-hdk_proc_macros = "=0.0.40-alpha1"
-holochain_wasm_utils = "=0.0.40-alpha1"
-holochain_json_derive = "=0.0.17"
-
-hdk_graph_helpers = { path = "../../../../../lib/hdk_graph_helpers" }
-vf_specification = { path = "../../../../../lib/vf_specification" }
-vf_observation = { path = "../../../../../lib/vf_observation" }
-vf_core = { path = "../../../../../lib/vf_core" }
-
-[lib]
-path = "src/lib.rs"
-crate-type = ["cdylib"]
