--- happs/specification_DEPRECATED/zomes/unit/code/.hcbuild
@@ -1,14 +1,0 @@
-{
-  "steps": [
-    {
-      "command": "cargo",
-      "arguments": [
-        "build",
-        "--release",
-        "--target=wasm32-unknown-unknown",
-        "--target-dir=/tmp/holochain/target"
-      ]
-    }
-  ],
-  "artifact": "/tmp/holochain/target/wasm32-unknown-unknown/release/hc_zome_rea_specification_unit.wasm"
-}
