+++ lib/rea_process_specification/defs/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "hc_zome_rea_process_specification_defs"
+version = "0.1.0"
+authors = ["pospi <pospi@spadgos.com>"]
+edition = "2018"
+
+[dependencies]
+serde = "1.0.89"
+serde_json = { version = "=1.0.39", features = ["preserve_order"] }
+serde_derive = "=1.0.89"
+# :DUPE: hdk-rust-revid
+hdk = "=0.0.40-alpha1"
+holochain_json_derive = "=0.0.17"
+
+hdk_graph_helpers = { path = "../../hdk_graph_helpers" }
+hc_zome_rea_process_specification_storage_consts = { path = "../storage_consts" }
+hc_zome_rea_process_specification_structs_internal = { path = "../structs_internal" }
+
+[lib]
+crate-type = ["lib"]
