+++ lib/rea_process_specification/lib/src/lib.rs
@@ -1,0 +1,112 @@
+/**
+ * Holo-REA process specification zome library API
+ *
+ * Contains helper methods that can be used to manipulate `ProcessSpecification` data
+ * structures in either the local Holochain zome, or a separate DNA-local zome.
+ *
+ * @package Holo-REA
+ */
+use std::borrow::Cow;
+use hdk::error::{ ZomeApiResult, ZomeApiError };
+
+use hdk_graph_helpers::{
+    records::{
+        create_record,
+        read_record_entry,
+        update_record,
+        delete_record,
+    },
+    links::{
+        get_linked_addresses_as_type,
+        get_linked_addresses_with_foreign_key_as_type,
+    },
+    local_indexes::{
+        query_direct_index_with_foreign_key,
+        query_direct_remote_index_with_foreign_key,
+    },
+    remote_indexes::{
+        RemoteEntryLinkResponse,
+        handle_sync_direct_remote_index_destination,
+    },
+};
+
+use hc_zome_rea_process_specification_storage_consts::*;
+use hc_zome_rea_process_specification_storage::*;
+use hc_zome_rea_process_specification_rpc::*;
+
+pub fn receive_create_process_specification(process_specification: CreateRequest) -> ZomeApiResult<Response> {
+    let (base_address, entry_resp): (ProcessSpecificationAddress, Entry) = create_record(
+        PROCESS_SPECIFICATION_BASE_ENTRY_TYPE,
+        PROCESS_SPECIFICATION_ENTRY_TYPE,
+        PROCESS_SPECIFICATION_INITIAL_ENTRY_LINK_TYPE,
+        process_specification.to_owned(),
+    )?;
+    Ok(construct_response(&base_address, &entry_resp))
+}
+pub fn receive_get_process_specification(address: ProcessSpecificationAddress) -> ZomeApiResult<Response> {
+    Ok(construct_response(&address, &read_record_entry(&address)?))
+}
+pub fn receive_update_process_specification(process_specification: UpdateRequest) -> ZomeApiResult<Response> {
+    handle_update_process_specification(&process_specification)
+}
+pub fn receive_delete_process_specification(address: ProcessSpecificationAddress) -> ZomeApiResult<bool> {
+    delete_record::<Entry>(&address)
+}
+pub fn receive_query_process_specifications(params: QueryParams) -> ZomeApiResult<Vec<Response>> {
+    handle_query_process_specifications(&params)
+}
+
+fn handle_update_process_specification(process_specification: &UpdateRequest) -> ZomeApiResult<Response> {
+    let address = process_specification.get_id();
+    let new_entry = update_record(PROCESS_SPECIFICATION_ENTRY_TYPE, &address, process_specification)?;
+    Ok(construct_response(address, &new_entry))
+}
+
+fn handle_query_process_specifications(_params: &QueryParams) -> ZomeApiResult<Vec<Response>> {
+    let entries_result: ZomeApiResult<Vec<(ProcessSpecificationAddress, Option<Entry>)>> = Err(ZomeApiError::Internal("No results found".to_string()));
+
+    // :TODO: implement "all" query and filters
+
+    match entries_result {
+        Ok(entries) => Ok(
+            entries.iter()
+                .map(|(entry_base_address, maybe_entry)| {
+                    match maybe_entry {
+                        Some(entry) => Ok(construct_response(
+                            entry_base_address,
+                            &entry,
+                        )),
+                        None => Err(ZomeApiError::Internal("referenced entry not found".to_string()))
+                    }
+                })
+                .filter_map(Result::ok)
+                .collect()
+        ),
+        Err(e) => Err(e)
+    }
+}
+
+
+/// Create response from input DHT primitives
+pub fn construct_response<'a>(
+    address: &ProcessSpecificationAddress, e: &Entry,
+) -> ResponseData {
+    ResponseData {
+        process_specification: Response {
+            id: address.to_owned(),
+            name: e.name.to_owned(),
+            note: e.note.to_owned(),
+        }
+    }
+}
+
+//---------------- READ ----------------
+
+// @see construct_response
+// pub fn get_link_fields<'a>(process_specification: &ProcessSpecificationAddress) -> (
+//     // :TODO:
+// ) {
+//     (
+//         // :TODO:
+//     )
+// }
