+++ lib/rea_process_specification/storage/src/lib.rs
@@ -1,0 +1,61 @@
+/**
+ * Holo-REA process specification zome internal data structures
+ *
+ * Required by the zome itself, and for any DNA-local zomes interacting with its
+ * storage API directly.
+ *
+ * @package Holo-REA
+ */
+extern crate serde;
+#[macro_use]
+extern crate serde_derive;
+extern crate serde_json;
+
+use holochain_json_api::{ json::JsonString, error::JsonError };
+use holochain_json_derive::{ DefaultJson };
+
+use hdk_graph_helpers::{
+    MaybeUndefined,
+    record_interface::Updateable,
+};
+
+use vf_core::type_aliases::{
+    Timestamp,
+    ExternalURL,
+    ProcessSpecificationAddress,
+    PlanAddress,
+};
+
+use hc_zome_rea_process_specification_rpc::{ CreateRequest, UpdateRequest };
+
+//---------------- RECORD INTERNALS & VALIDATION ----------------
+
+#[derive(Serialize, Deserialize, Debug, DefaultJson, Default, Clone)]
+pub struct Entry {
+    pub name: String,
+    pub note: Option<String>,
+}
+
+//---------------- CREATE ----------------
+
+/// Pick relevant fields out of I/O record into underlying DHT entry
+impl From<CreateRequest> for Entry {
+    fn from(e: CreateRequest) -> Entry {
+        Entry {
+            name: e.name.into(),
+            note: e.note.into(),
+        }
+    }
+}
+
+//---------------- UPDATE ----------------
+
+/// Handles update operations by merging any newly provided fields
+impl Updateable<UpdateRequest> for Entry {
+    fn update_with(&self, e: &UpdateRequest) -> Entry {
+        Entry {
+            name: if !e.name.is_some() { self.name.to_owned() } else { e.name.to_owned().unwrap() },
+            note: if e.note.is_undefined() { self.note.to_owned() } else { e.note.to_owned().into() },
+        }
+    }
+}
