+++ lib/rea_process_specification/storage_consts/src/lib.rs
@@ -1,0 +1,10 @@
+/**
+ * Storage constants for zome entry & link type identifiers
+ *
+ * Used by modules interfacing with the underlying Holochain storage system directly.
+ *
+ * @package Holo-REA
+ */
+pub const PROCESS_SPECIFICATION_ENTRY_TYPE: &str = "vf_process_specification";
+pub const PROCESS_SPECIFICATION_BASE_ENTRY_TYPE: &str = "vf_process_specification_baseurl";
+pub const PROCESS_SPECIFICATION_INITIAL_ENTRY_LINK_TYPE: &str = "vf_process_specification_entry";
