+++ lib/rea_resource_specification/lib/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "hc_zome_rea_resource_specification_lib"
+version = "0.1.0"
+authors = ["pospi <pospi@spadgos.com>"]
+edition = "2018"
+
+[dependencies]
+# :DUPE: hdk-rust-revid
+hdk = "=0.0.40-alpha1"
+
+hdk_graph_helpers = { path = "../../hdk_graph_helpers" }
+vf_core = { path = "../../vf_core" }
+hc_zome_rea_resource_specification_storage_consts = { path = "../storage_consts" }
+hc_zome_rea_resource_specification_storage = { path = "../storage" }
+hc_zome_rea_resource_specification_rpc = { path = "../rpc" }
+hc_zome_TODO_storage_consts = { path = "../../TODO/storage_consts" }
+
+[lib]
+crate-type = ["lib"]
