+++ lib/rea_resource_specification/storage/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "hc_zome_rea_resource_specification_storage"
+version = "0.1.0"
+authors = ["pospi <pospi@spadgos.com>"]
+edition = "2018"
+
+[dependencies]
+serde = "1.0.89"
+serde_json = { version = "=1.0.39", features = ["preserve_order"] }
+serde_derive = "=1.0.89"
+# :DUPE: hdk-rust-revid
+holochain_json_api = "=0.0.17"
+holochain_json_derive = "=0.0.17"
+
+hdk_graph_helpers = { path = "../../hdk_graph_helpers" }
+vf_core = { path = "../../vf_core" }
+hc_zome_rea_resource_specification_rpc = { path = "../structs_rpc" }
+
+[lib]
+crate-type = ["lib"]
