+++ lib/rea_resource_specification/storage_consts/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "hc_zome_rea_resource_specification_storage_consts"
+version = "0.1.0"
+authors = ["pospi <pospi@spadgos.com>"]
+edition = "2018"
+
+[dependencies]
+
+[lib]
+crate-type = ["lib"]
