+++ lib/rea_resource_specification/storage_consts/src/lib.rs
@@ -1,0 +1,12 @@
+/**
+ * Storage constants for zome entry & link type identifiers
+ *
+ * Used by modules interfacing with the underlying Holochain storage system directly.
+ *
+ * @package Holo-REA
+ */
+pub const ECONOMIC_RESOURCE_SPECIFICATION_ENTRY_TYPE: &str = "vf_resource_specification";
+pub const ECONOMIC_RESOURCE_SPECIFICATION_BASE_ENTRY_TYPE: &str = "vf_resource_specification_baseurl";
+pub const ECONOMIC_RESOURCE_SPECIFICATION_INITIAL_ENTRY_LINK_TYPE: &str = "vf_resource_specification_entry";
+pub const RESOURCE_SPECIFICATION_CONFORMING_RESOURCE_LINK_TYPE: &str = "vf_resource_specification_conforming_resource";
+pub const RESOURCE_SPECIFICATION_CONFORMING_RESOURCE_LINK_TAG: &str = "conforming_resource";
