+++ lib/rea_unit/storage_consts/src/lib.rs
@@ -1,0 +1,10 @@
+/**
+ * Storage constants for zome entry & link type identifiers
+ *
+ * Used by modules interfacing with the underlying Holochain storage system directly.
+ *
+ * @package Holo-REA
+ */
+pub const UNIT_ENTRY_TYPE: &str = "vf_unit";
+pub const UNIT_ID_ENTRY_TYPE: &str = "vf_unit_id";
+pub const UNIT_INITIAL_ENTRY_LINK_TYPE: &str = "vf_unit_entry";
