--- lib/vf_specification/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "vf_specification"
-version = "0.1.0"
-authors = ["JhonatanHern <jhonatanhernandez998@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-serde = "=1.0.89"
-serde_json = { version = "=1.0.39", features = ["preserve_order"] }
-serde_derive = "=1.0.89"
-# :DUPE: hdk-rust-revid
-hdk = "=0.0.40-alpha1"
-holochain_json_derive = "=0.0.17"
-
-hdk_graph_helpers = { path = "../hdk_graph_helpers" }
-vf_core = { path = "../vf_core" }
-vf_knowledge = { path = "../vf_knowledge" }
-
-[lib]
-crate-type = ["lib"]
